Initial version of board-specific init.rc
The base init.rc from system/core is now used as-is.
init.am335xevm_sk.rc is renamed to init.am335xevm.rc in the filesystem
as {ro.hardware} shows up as am335xevm and not am335xevm_sk
since we are using the same kernel source as am335xevm for am335xevm-sk too.

Change-Id: I416d5e2ec5db191b0f312ab27e8bb9104045bbc3
Signed-off-by: Vishveshwar Bhat <vishveshwar.bhat@ti.com>